+++ Microsoft.Azure.Cosmos/src/Routing/GatewayAddressCache.cs
@@ -357,7 +357,6 @@ namespace Microsoft.Azure.Cosmos.Routing
 
             string resourceTypeToSign = PathsHelper.GetResourcePath(resourceType);
 
-            string payload;
             headers.Set(HttpConstants.HttpHeaders.XDate, DateTime.UtcNow.ToString("r", CultureInfo.InvariantCulture));
             string token = this.tokenProvider.GetUserAuthorizationToken(
                 resourceAddress,
@@ -365,7 +364,7 @@ namespace Microsoft.Azure.Cosmos.Routing
                 HttpConstants.HttpMethods.Get,
                 headers,
                 AuthorizationTokenType.PrimaryMasterKey,
-                out payload);
+                payload: out _);
 
             headers.Set(HttpConstants.HttpHeaders.Authorization, token);
 
@@ -414,14 +413,13 @@ namespace Microsoft.Azure.Cosmos.Routing
             string token = null;
             try
             {
-                string payload;
                 token = this.tokenProvider.GetUserAuthorizationToken(
                     collectionRid,
                     resourceTypeToSign,
                     HttpConstants.HttpMethods.Get,
                     headers,
                     AuthorizationTokenType.PrimaryMasterKey,
-                    out payload);
+                    payload: out _);
             }
             catch (UnauthorizedException)
             {
@@ -430,7 +428,6 @@ namespace Microsoft.Azure.Cosmos.Routing
             if (token == null && request.IsNameBased)
             {
                 // User doesn't have rid based resource token. Maybe he has name based.
-                string payload;
                 string collectionAltLink = PathsHelper.GetCollectionPath(request.ResourceAddress);
                 token = this.tokenProvider.GetUserAuthorizationToken(
                         collectionAltLink,
@@ -438,7 +435,7 @@ namespace Microsoft.Azure.Cosmos.Routing
                         HttpConstants.HttpMethods.Get,
                         headers,
                         AuthorizationTokenType.PrimaryMasterKey,
-                        out payload);
+                        payload: out _);
             }
 
             headers.Set(HttpConstants.HttpHeaders.Authorization, token);