+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/CosmosClientOptionsUnitTests.cs
@@ -79,8 +79,9 @@ namespace Microsoft.Azure.Cosmos.Tests
             Assert.IsNull(policy.OpenTcpConnectionTimeout);
             Assert.IsNull(policy.MaxRequestsPerTcpConnection);
             Assert.IsNull(policy.MaxTcpConnectionsPerEndpoint);
+            Assert.IsTrue(policy.EnableEndpointDiscovery);
 
-            cosmosClientBuilder.WithApplicationRegion(region)
+            cosmosClientBuilder.WithApplicationRegion(region, false)
                 .WithConnectionModeGateway(maxConnections, webProxy)
                 .WithRequestTimeout(requestTimeout)
                 .WithApplicationName(userAgentSuffix)
@@ -108,6 +109,7 @@ namespace Microsoft.Azure.Cosmos.Tests
             Assert.AreEqual(cosmosSerializerOptions.Indented, clientOptions.SerializerOptions.Indented);
             Assert.IsTrue(object.ReferenceEquals(webProxy, clientOptions.WebProxy));
             Assert.IsTrue(clientOptions.AllowBulkExecution);
+            Assert.IsFalse(clientOptions.UseAnyAccountRegion);
 
             //Verify GetConnectionPolicy returns the correct values
             policy = clientOptions.GetConnectionPolicy();
@@ -118,6 +120,7 @@ namespace Microsoft.Azure.Cosmos.Tests
             Assert.AreEqual(requestTimeout, policy.RequestTimeout);
             Assert.IsTrue(policy.UserAgentSuffix.Contains(userAgentSuffix));
             Assert.IsTrue(policy.UseMultipleWriteLocations);
+            Assert.IsFalse(policy.EnableEndpointDiscovery);
             Assert.AreEqual(maxRetryAttemptsOnThrottledRequests, policy.RetryOptions.MaxRetryAttemptsOnThrottledRequests);
             Assert.AreEqual((int)maxRetryWaitTime.TotalSeconds, policy.RetryOptions.MaxRetryWaitTimeInSeconds);
 
@@ -360,6 +363,16 @@ namespace Microsoft.Azure.Cosmos.Tests
             HttpClientHandler innerHandler = (HttpClientHandler)handler.InnerHandler;
 
             Assert.IsTrue(object.ReferenceEquals(webProxy, innerHandler.Proxy));
+        }
+
+        [TestMethod]
+        public void VerifyCorrectProtocolIsSet()
+        {
+            CosmosClientOptions cosmosClientOptions = new CosmosClientOptions { ConnectionMode = ConnectionMode.Gateway };
+            Assert.AreEqual(Protocol.Https, cosmosClientOptions.ConnectionProtocol);
+
+            cosmosClientOptions = new CosmosClientOptions { ConnectionMode = ConnectionMode.Direct };
+            Assert.AreEqual(Protocol.Tcp, cosmosClientOptions.ConnectionProtocol);
         }
 
         private class TestWebProxy : IWebProxy